Funcionando con la comparacion de un registro
