+++ .settings/javaswv.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<JavaSwv>
+<DwtPcSampleRate CycleTap="false" PostScaler="8"/>
+<config clockSpeed="96000000"/>
+</JavaSwv>
