+++ ControlVehicular LinkServer Debug.launch
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="com.crt.dsfdebug.crtmcu.launchType">
+<stringAttribute key="InstructionTraceInfo" value="eJytVltvmzAUfp+0/4B47wIhoVRqVjngLGjcZJu22gtyiZOyEsi4dN2/n7nkTi6alofIPv7Od64+5v7hYxkL7yzLozQZifIXSRRYEqazKFmMRJ9MbjTx4evnT/esWLLixWAFjeK8EhCaLVgxYbQoM5YLNv2ZZo9rHs5iR8m+oFHgpA72Pc9FBBpiRQSJPaY5E8BsxonykXgzVFRVGqi3fVHwuIglXEsRexXYeCLHYE1VNVndARdZyRq82UWu3Up3cr8TD8m4yxlpeLvLP6dx3irYRwryCSDP3u/iANtXtYEiK6rUrUMyGrJJmSQs3vdoeKcONKkzAm5ETzOGo8UrXyxXacIBQmW3M2d6bqXh20h0yjhu1qLAKRBb4ILT8MpJVBsqoSwNpc5fKKmhTIfs1Hn1U+SZJs1PHIaSptxqs0MxVcW24PN1j1VB7HnfAkIeJs0o9zcXnHL5wrKROGiygcvVKs0KNhP05cqmRfja5krQ/4Qx09MyKSp4LasUqpS1XS4YtKDVfl0TwYqSN3mzs+mHTXOeO3kobuxKjd1umsZyy9JsOknkcyTX+9L/PzTt1ettctm2mZel8yjmg6LNYt6WI9AJsoRHGpdsk4/eKXxvp8Br4HH/1lezWF7T2OvJsW7spzSbbRqbvZQLxH5t4uWcuKBZwXt9VQ8pvovjzXG9s9g7i5sJFi1ZXtCd7NUXFGw12nk22Qlo24GmMRL5oWmgwEMQQ4cc9miFIAjoMMAEIML/Xe8cdgpwMDEn7sS3rK5zYBGIHEDMRxiMEXD0aeAB/TvsNjz17XGg9IN2PHcTWgGC30zMeTFfAQOMLdiW+CDwusoHKahlVX83ybDH0DCgEZg84idA9Knnmg4JTMfzCT66wI2u1Oo+E+hUutWCx2g1WgGGFtSJi/5RfezjAJs/4HltD5goAIbBK4MD3bU9gMBlmwYg4Ap0f9DAq8IGBvTI9BRwG8o2ggs+6K7vVLW7hOKgZ3J9aBsfXJ+ccaKNzHQwQb7N+we7PtLhBXLbNqDuGhDtRdjb9levczTUz/91M2P9QdAxMw7f4SM61sG3fp47Htf6ZmAyOZyP5000kL0vrlq0/1X2F2z9+TYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"/>
 <stringAttribute key="com.crt.ctrlcenter.OFSemuDetails" value="LinkServer"/>
 <booleanAttribute key="com.crt.ctrlcenter.crtInit" value="true"/>
 <stringAttribute key="com.crt.ctrlcenter.currentWireType" value="SWD"/>
