+++ Debug/Example/src/RFID.d
@@ -1,4 +1,4 @@
-Example/src/FRTOS.o Example/src/FRTOS.d: ../Example/src/FRTOS.c \
+Example/src/RFID.o Example/src/RFID.d: ../Example/src/RFID.c \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/chip.h \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/lpc_types.h \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/sys_config.h \
@@ -47,9 +47,7 @@ Example/src/FRTOS.o Example/src/FRTOS.d: ../Example/src/FRTOS.c \
  C:\Users\kevin\git\ControlVehicular\freeRTOS\inc/list.h \
  C:\Users\kevin\git\ControlVehicular\freeRTOS\inc/semphr.h \
  C:\Users\kevin\git\ControlVehicular\freeRTOS\inc/queue.h \
- C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h \
  C:\Users\kevin\git\ControlVehicular\Example\inc/MFRC522.h \
- C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h \
  C:\Users\kevin\git\ControlVehicular\Example\inc/rfid_utils.h \
  C:\Users\kevin\git\ControlVehicular\Example\inc/MFRC522.h
 
@@ -149,11 +147,7 @@ C:\Users\kevin\git\ControlVehicular\freeRTOS\inc/semphr.h:
 
 C:\Users\kevin\git\ControlVehicular\freeRTOS\inc/queue.h:
 
-C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h:
-
 C:\Users\kevin\git\ControlVehicular\Example\inc/MFRC522.h:
-
-C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h:
 
 C:\Users\kevin\git\ControlVehicular\Example\inc/rfid_utils.h:
