+++ Debug/Example/src/rfid_utils.d
@@ -2,7 +2,6 @@ Example/src/rfid_utils.o Example/src/rfid_utils.d: \
  ../Example/src/rfid_utils.c \
  C:\Users\kevin\git\ControlVehicular\Example\inc/rfid_utils.h \
  C:\Users\kevin\git\ControlVehicular\Example\inc/MFRC522.h \
- C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/chip.h \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/lpc_types.h \
  D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/sys_config.h \
@@ -55,8 +54,6 @@ Example/src/rfid_utils.o Example/src/rfid_utils.d: \
 C:\Users\kevin\git\ControlVehicular\Example\inc/rfid_utils.h:
 
 C:\Users\kevin\git\ControlVehicular\Example\inc/MFRC522.h:
-
-C:\Users\kevin\git\ControlVehicular\Example\inc/delay.h:
 
 D:\Facultad\Digitales\ II\Repositorio\ MCU\lpc_chip_175x_6x\inc/chip.h:
