+++ Debug/Example/src/subdir.mk
@@ -4,24 +4,24 @@
 
 # Add inputs and outputs from these tool invocations to the build variables 
 C_SRCS += \
-../Example/src/FRTOS.c \
 ../Example/src/MFRC522.c \
+../Example/src/RFID.c \
 ../Example/src/cr_startup_lpc175x_6x.c \
 ../Example/src/crp.c \
 ../Example/src/rfid_utils.c \
 ../Example/src/sysinit.c 
 
 OBJS += \
-./Example/src/FRTOS.o \
 ./Example/src/MFRC522.o \
+./Example/src/RFID.o \
 ./Example/src/cr_startup_lpc175x_6x.o \
 ./Example/src/crp.o \
 ./Example/src/rfid_utils.o \
 ./Example/src/sysinit.o 
 
 C_DEPS += \
-./Example/src/FRTOS.d \
 ./Example/src/MFRC522.d \
+./Example/src/RFID.d \
 ./Example/src/cr_startup_lpc175x_6x.d \
 ./Example/src/crp.d \
 ./Example/src/rfid_utils.d \
