+++ Example/src/RFID.c
@@ -1,79 +1,55 @@
 /*
- * RFID.c
- *
- *  Created on: Aug 2, 2018
- *      Author: kevin
- */
-
-
+===============================================================================
+ Name        : FRTOS.c
+ Author      : $(author)
+ Version     :
+ Copyright   : $(copyright)
+ Description : main definition
+===============================================================================
+*/
 #include "chip.h"
-#include <cr_section_macros.h>
-/*******************************************************************************
- * Custom files and libraries includes
- ******************************************************************************/
-#include "delay.h"
-#include "MFRC522.h"
-#include "rfid_utils.h"
 #include "FreeRTOS.h"
 #include "task.h"
 #include "semphr.h"
+#include "MFRC522.h"
+#include "rfid_utils.h"
 
-#define MIN_BALANCE 300
 
+
+//#include "RegsLPC1769.h"
+
+
+
+#define DEBUGOUT1(...) printf(__VA_ARGS__)
 #define DEBUGOUT(...) printf(__VA_ARGS__)
 #define DEBUGSTR(...) printf(__VA_ARGS__)
 
-/*******************************************************************************
- *  Extra functions defined in the main.c file
- ******************************************************************************/
-void userTapIn();
+#define MIN_BALANCE 300
 
-/*******************************************************************************
+
+/*****************************************************************************
  * Private types/enumerations/variables
- ******************************************************************************/
-int last_balance = 0;
-unsigned int last_user_ID;
+ ****************************************************************************/
 
-/*******************************************************************************
- * Public types/enumerations/variables
- ******************************************************************************/
+static volatile bool fIntervalReached;
+static volatile bool fAlarmTimeMatched;
+static volatile bool On0, On1;
 
 // RFID structs
 MFRC522Ptr_t mfrcInstance;
 
+int last_balance = 0;
+unsigned int last_user_ID;
 
-/**
- * @brief	Main program body
- * @return	Does not return
- */
-int main(void) {
-	// Read clock settings and update SystemCoreClock variable
-	SystemCoreClockUpdate();
+/*****************************************************************************
+ * Public types/enumerations/variables
+ ****************************************************************************/
 
-	/* Sets up DEBUG UART */
-//	DEBUGINIT();
+SemaphoreHandle_t Semaforo_RFID;
 
-	/* Initializes GPIO */
-	Chip_GPIO_Init(LPC_GPIO);
-
-	//Voy a utilizar la SSP1
-	setupRFID(&mfrcInstance);
-
-	while (1) {
-
-		// Look for new cards in RFID2
-		if (PICC_IsNewCardPresent(mfrcInstance)) {
-			// Select one of the cards
-			if (PICC_ReadCardSerial(mfrcInstance)) {
-//				int status = writeCardBalance(mfrcInstance, 100000); // used to recharge the card
-				 userTapIn();
-			}
-		}
-
-		__WFI();
-	}
-}
-
+/*****************************************************************************
+ * Private functions
+ ****************************************************************************/
 /*******************************************************************************
  *  System routine functions
  ******************************************************************************/
@@ -81,7 +57,9 @@ int main(void) {
 /**
  * Executed every time the card reader detects a user in
  */
-void userTapIn() {
+void userTapIn()
+
+{
 
 //	show card UID
 	DEBUGOUT("\nCard uid bytes: ");
@@ -98,21 +76,81 @@ void userTapIn() {
 		(int)mfrcInstance->uid.uidByte[1] << 16 |
 		(int)mfrcInstance->uid.uidByte[0] << 24;
 
-	DEBUGOUT("\nCard Read user ID: %u ",last_user_ID);
+	DEBUGOUT("\nCard Read user ID: %u \n\r",last_user_ID);
 
+	Comparar(last_user_ID);
 
-		// Read the user balance
-		last_balance = readCardBalance(mfrcInstance);
+	// Read the user balance NO BORRAR SINO NO DETECTA CUANDO HAY TARJETA NUEVA
+	last_balance = readCardBalance(mfrcInstance);
 
-		if (last_balance == (-999)) {
-			// Error handling, the card does not have proper balance data inside
-		} else {
-			// Check for minimum balance
-			if (last_balance < MIN_BALANCE) {
-				DEBUGOUT(", Insufficient balance\n");
-			} else {
-				DEBUGOUT(", Balance OK\n");
-			}
-		}
 }
 
+/*****************************************************************************
+ * Public functions
+ ****************************************************************************/
+
+
+static void xTaskRFIDConfig(void *pvParameters)
+{
+	SystemCoreClockUpdate();
+
+	DEBUGOUT1("PRUEBA RFID..\n");	//Imprimo en la consola
+
+	setupRFID(&mfrcInstance);
+
+	xSemaphoreGive(Semaforo_RFID);//Doy el semaforo para que se inicie la lectura
+
+	vTaskDelete(NULL);	//Borra la tarea
+}
+
+///////////////////////////////////////////////////////////////////////////////////////////////////////////
+/* vTaskInicTimer */
+static void vTaskRFID(void *pvParameters)
+{
+	xSemaphoreTake(Semaforo_RFID, portMAX_DELAY); //semaforo de inicializacion de rfid
+	while (1)
+	{
+
+		// Look for new cards in RFID2
+		if (PICC_IsNewCardPresent(mfrcInstance))
+		{
+			// Select one of the cards
+			if (PICC_ReadCardSerial(mfrcInstance))
+			{
+//				int status = writeCardBalance(mfrcInstance, 100000); // used to recharge the card
+				 userTapIn();
+			}
+		}
+		vTaskDelay( 1000 / portTICK_PERIOD_MS );//Muestreo cada 1 seg
+	}
+
+	vTaskDelete(NULL);	//Borra la tarea si sale del while
+}
+
+///////////////////////////////////////////////////////////////////////////////////////////////////////////
+/* main
+*/
+int main(void)
+{
+	SystemCoreClockUpdate();
+
+	/* Initializes GPIO */
+	Chip_GPIO_Init(LPC_GPIO);
+
+	vSemaphoreCreateBinary(Semaforo_RFID);
+	xSemaphoreTake(Semaforo_RFID, portMAX_DELAY); //semaforo de inicializacion de rfid
+
+	xTaskCreate(vTaskRFID, (char *) "vTaskRFID",
+				configMINIMAL_STACK_SIZE, NULL, (tskIDLE_PRIORITY + 1UL),
+				(xTaskHandle *) NULL);
+	xTaskCreate(xTaskRFIDConfig, (char *) "xTaskRFIDConfig",
+				configMINIMAL_STACK_SIZE, NULL, (tskIDLE_PRIORITY + 2UL),
+				(xTaskHandle *) NULL);
+
+
+	/* Start the scheduler */
+	vTaskStartScheduler();
+
+	/* Nunca debería arribar aquí */
+    return 0;
+}
